Rearranged some DevTools Components code
