+++ packages/react-devtools-shared/src/devtools/views/Components/Components.css
@@ -1,3 +1,14 @@
+.Components {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: row;
+  background-color: var(--color-background);
+  color: var(--color-text);
+  font-family: var(--font-family-sans);
+}
+
 .TreeWrapper {
   flex: 0 0 var(--horizontal-resize-percentage);
   overflow: auto;
@@ -23,6 +34,10 @@
 }
 
 @media screen and (max-width: 600px) {
+  .Components {
+    flex-direction: column;
+  }
+
   .TreeWrapper {
     flex: 0 0 var(--vertical-resize-percentage);
   }
