--- packages/react-devtools-shared/src/devtools/views/Components/ComponentsResizer.css
@@ -1,16 +1,0 @@
-.ComponentsWrapper {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  background-color: var(--color-background);
-  color: var(--color-text);
-  font-family: var(--font-family-sans);
-}
-
-@media screen and (max-width: 600px) {
-  .ComponentsWrapper {
-    flex-direction: column;
-  }
-}
